+++ config/filament-trumbowyg.php
@@ -2,4 +2,5 @@
 
 return [
     'buttons' => [],
+    'tagClasses' => []
 ];