+++ config/filament-trumbowyg.php
@@ -6,4 +6,5 @@ return [
     'change_active_dropdown_icon' => false,
     'semantic' => true,
     'remove_format_pasted' => false,
+    'tags_to_remove' => [],
 ];