+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to `filament-trumbowyg` will be documented in this file.
 - Adds support for active dropdown icon [https://alex-d.github.io/Trumbowyg/documentation/#change-active-dropdown-icon]
 - Adds support for remove format on pasted [https://alex-d.github.io/Trumbowyg/documentation/#remove-format-pasted]
 - Adds support for removing tags [https://alex-d.github.io/Trumbowyg/documentation/#tags-to-remove]
+- Adds support for keeping tags [https://alex-d.github.io/Trumbowyg/documentation/#tags-to-keep]
 
 ## 2.0.0 - 2023-04-12
 - Adds support for Laravel 10